added another leetcode solution
Not the most optimal solution. Too many lines used. But at this point of time, it is the most efficient method i can come up with. Much better than using loop and check if letters exists etc. This check if certain letter combination exists but only once in a while statement, especially if the letter combination doesn't exists etc IV